Adding hw-offload solution to OVS using userspace datapath and DPDK
This solution use the generic flow API in DPDK -RTE FLOW- to automate
flow classification offload to NIC.

Since OVS 2.7.0 release does only support DPDK 16.11, this commit
contains a quick and dirty upgrade to 17.02. This commit solely
serves the purpose to initiate the first attempt to fulfill the
hw-offload feature, using DPDK.

It must be merged into master for a final patch proposal, once OVS is
upgraded to run on a newer DPDK with RTE FLOW support.

When enabled on a port it tries to hw-offload flow entries in the flow
cache on add, modify and delete operations. On success, a flow id is
added to mbuf header on hit to indicate pre-classified on receival.

Once OVS receives a pre-classified packet it directs it directly to
output. Flow statistics are either retreived by NIC or handled in OVS.

This is the first step of a solution proposal, which currently only
supports north traffic acceleration. Further steps are planned to
handle more directions to accelerate. North,south,east and west. This
requires virtual port concepts in NICs.