Use netdev flow_get function to add support for RTE FLOW query
To be able to query NIC for flow statistics of the fully offloaded flows,
the flow_get function is used to call RTE FLOW query function in netdev-dpdk.

The statistics returned by full offloaded flows is used to update the mega-flow
cache flow statistics, for seemless integration and easy SW failover.